Renamed of the Firehose service gRPC ids and packages (#3054)
This will require a coordinate update for NEAR when Firehose service ids will change there, `graph-node` with this version will need to be deployed.
