+++ chain/ethereum/proto/codec.proto
@@ -1,6 +1,8 @@
 syntax = "proto3";
 
-package dfuse.ethereum.codec.v1;
+package sf.ethereum.codec.v1;
+
+option go_package = "github.com/streamingfast/sf-ethereum/pb/sf/ethereum/codec/v1;pbcodec";
 
 import "google/protobuf/timestamp.proto";
