+++ graph/build.rs
@@ -3,6 +3,6 @@ fn main() {
     tonic_build::configure()
         .out_dir("src/firehose")
         .format(true)
-        .compile(&["proto/bstream.proto"], &["proto"])
+        .compile(&["proto/firehose.proto"], &["proto"])
         .expect("Failed to compile Firehose proto(s)");
 }
