+++ graph/src/firehose/codec.rs
@@ -1,0 +1,4 @@
+#[path = "sf.firehose.v1.rs"]
+mod pbfirehose;
+
+pub use pbfirehose::*;
